src/main/java/de/tudarmstadt/aiphes/mdswriter/data/NuggetGroupManager.java
/*******************************************************************************
 * Copyright 2016
 * Ubiquitous Knowledge Processing (UKP) Lab
 * Technische Universität Darmstadt
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 ******************************************************************************/
package de.tudarmstadt.aiphes.mdswriter.data;

import java.sql.PreparedStatement;
import java.sql.SQLException;
import java.util.ArrayList;
import java.util.HashSet;
import java.util.List;
import java.util.Set;

import javax.websocket.Session;

import de.tudarmstadt.aiphes.mdswriter.auth.AuthManager;
import de.tudarmstadt.aiphes.mdswriter.auth.User;
import de.tudarmstadt.aiphes.mdswriter.db.DBConnection;
import de.tudarmstadt.aiphes.mdswriter.db.DBManager;
import de.tudarmstadt.aiphes.mdswriter.doc.DocumentManager;
import de.tudarmstadt.aiphes.mdswriter.doc.DocumentSet;

public class NuggetGroupManager extends NuggetManager<NuggetGroupItem> {

	public NuggetGroupManager(final User user, final DocumentSet docSet)
			throws SQLException {
		super(user, docSet);

		// Load the nuggets ordered by orderPos.
		DBConnection connection = DBManager.getConnection();
		try {
			nuggets = Nugget.list(connection, user, docSet);
			createItemList();
		} finally {
			connection.close();
		}
	}

	protected NuggetGroupManager(final List<Nugget> nuggets) {
		super(nuggets);
	}

	protected void createItemList() {
		items = new ArrayList<NuggetGroupItem>();
		int prevGroup = 0;
		int newGroup = 0;
		for (Nugget nugget : nuggets) {
			if (nugget.getGroup() != prevGroup) {
				newGroup++;
				items.add(createGroupItem(newGroup, null));
				prevGroup = nugget.getGroup();
			}
			items.add(createNuggetItem(nugget, newGroup, null));
		}
	}

	protected NuggetGroupItem createGroupItem(final int groupId,
			final String groupTitle) {
		return new NuggetGroupItem(groupId);
	}

	protected NuggetGroupItem createNuggetItem(final Nugget nugget,
			final int groupId, final String groupTitle) {
		return new NuggetGroupItem(nugget, groupId);
	}

	protected void doSaveItems(final DBConnection connection,
			final Set<NuggetGroupItem> itemsToSave) throws SQLException {
		if (itemsToSave == null || itemsToSave.isEmpty())
			return;

		if (DEBUG_MODE)
			System.out.println("SAVING TO DB!");
		PreparedStatement stmt = connection.prepareStatement(
				"UPDATE " + Nugget.table().getName()
				+ " SET nugget_group = ?, nugget_orderpos = ?, nugget_isbest = ?"
				+ " WHERE id = ?");
		try {
			for (NuggetGroupItem item : itemsToSave)
				item.save(stmt);
		} finally {
			stmt.close();
		}
	}

	public void selectBestNugget(int groupId, int nuggetId) throws SQLException {
		saveItems(doSelectBestNugget(groupId, nuggetId));
	}

	public void deselectBestNugget(int groupId) throws SQLException {
		saveItems(doSelectBestNugget(groupId, -1));
	}

	protected Set<NuggetGroupItem> doSelectBestNugget(int groupId, int nuggetId) {
		Set<NuggetGroupItem> itemsToSave = new HashSet<NuggetGroupItem>();
		boolean foundGroup = false;
		for (NuggetGroupItem item : items) {
			if (item.getGroupId() == groupId) {
				if (!item.isGroupHeader()) {
					if (item.setBestNugget(item.getNugget().getId() == nuggetId))
						itemsToSave.add(item);
				}
				foundGroup = true;
			} else
			if (foundGroup)
				break;
		}
		return itemsToSave;
	}

	protected void dumpItems() {
		System.out.println();
		int j = 0;
		for (NuggetGroupItem i : items) {
			System.out.printf("%4d  ", j);
			if (i.isGroupHeader())
				System.out.printf("[---] Group %3d\n",
						i.getGroupId());
			else
				System.out.printf("[%3d] Group %3d > %3d  /  OrderPos %3d > %3d / Best: %1d\n",
						i.getNugget().getId(),
						i.getGroupId(), i.getNugget().getGroup(),
						i.getNugget().getOrderPos(), i.computeNewPos(j),
						i.getNugget().isBestNugget() ? 1 : 0);
			j++;
		}
		System.out.println();
	}

	protected void dumpNuggets() {
		for (Nugget n : nuggets)
			System.out.println(n.getId() + " " + n.getGroup() + " " + n.getOrderPos());
		System.out.println();
	}


	public static NuggetGroupManager get(final Session session)
			throws SQLException {
		NuggetGroupManager result = (NuggetGroupManager) session.getUserProperties().get("nuggets");
		if (result == null) {
			DocumentSet docSet = DocumentManager.getSessionDocSet(session);
			if (docSet == null)
				throw new NullPointerException("Session docSet missing");

			User user = AuthManager.getAccessToken(session).getUser();
			result = new NuggetGroupManager(user, docSet);
			session.getUserProperties().put("nuggets", result);
		}
		return result;
	}

}
