src/test/java/de/tudarmstadt/aiphes/mdswriter/data/NuggetGroupManagerTest.java
/*******************************************************************************
 * Copyright 2016
 * Ubiquitous Knowledge Processing (UKP) Lab
 * Technische Universität Darmstadt
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 ******************************************************************************/
package de.tudarmstadt.aiphes.mdswriter.data;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.Iterator;
import java.util.List;
import java.util.Set;
import java.util.TreeSet;

import de.tudarmstadt.aiphes.mdswriter.data.Nugget;
import de.tudarmstadt.aiphes.mdswriter.data.NuggetGroupItem;
import de.tudarmstadt.aiphes.mdswriter.data.NuggetGroupManager;
import junit.framework.TestCase;

public class NuggetGroupManagerTest extends TestCase {

	protected static class TestNugget extends Nugget {

		public TestNugget(final int id, final int group, final int orderPos) {
			super(id, null, null, null, "", null, -1, -1, group, orderPos, false, "", "", null);
		}

		public TestNugget(final int id, final int group, final int orderPos, final boolean bestNugget) {
			super(id, null, null, null, "", null, -1, -1, group, orderPos, bestNugget, "", "", null);
		}

	}

	public void testMakeConsistentGroups() {
		List<Nugget> nuggets = new ArrayList<Nugget>();
		nuggets.add(new TestNugget(101, 0, 0));
		nuggets.add(new TestNugget(102, 0, 0));
		nuggets.add(new TestNugget(103, 0, 0));
		nuggets.add(new TestNugget(104, 0, 0));
		nuggets.add(new TestNugget(105, 0, 0));

		NuggetGroupManager nm = new NuggetGroupManager(nuggets);
		Set<NuggetGroupItem> itemsToSave = nm.doMakeConsistent();
		Iterator<Nugget> iter = nm.getNuggets().iterator();
		assertNugget(101, 0, 1, iter.next());
		assertNugget(102, 0, 2, iter.next());
		assertNugget(103, 0, 3, iter.next());
		assertNugget(104, 0, 4, iter.next());
		assertNugget(105, 0, 5, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 101, 102, 103, 104, 105);

		itemsToSave = nm.doMakeConsistent();
		iter = nm.getNuggets().iterator();
		assertNugget(101, 0, 1, iter.next());
		assertNugget(102, 0, 2, iter.next());
		assertNugget(103, 0, 3, iter.next());
		assertNugget(104, 0, 4, iter.next());
		assertNugget(105, 0, 5, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave);


		nuggets = new ArrayList<Nugget>();
		nuggets.add(new TestNugget(101, 2, 2));
		nuggets.add(new TestNugget(102, 1, 3));
		nuggets.add(new TestNugget(103, 1, 5));
		nuggets.add(new TestNugget(104, 1, 5));
		nuggets.add(new TestNugget(105, 3, 5));
		nuggets.add(new TestNugget(106, 1, 9));

		nm = new NuggetGroupManager(nuggets);
		itemsToSave = nm.doMakeConsistent();
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, 1, iter.next());
		assertNugget(102, 2, 2, iter.next());
		assertNugget(103, 2, 3, iter.next());
		assertNugget(104, 2, 4, iter.next());
		assertNugget(105, 3, 5, iter.next());
		assertNugget(106, 4, 6, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 101, 102, 103, 104, 106);

		itemsToSave = nm.doMakeConsistent();
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, 1, iter.next());
		assertNugget(102, 2, 2, iter.next());
		assertNugget(103, 2, 3, iter.next());
		assertNugget(104, 2, 4, iter.next());
		assertNugget(105, 3, 5, iter.next());
		assertNugget(106, 4, 6, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave);
	}

	public void testAddGroup() {
		List<Nugget> nuggets = new ArrayList<Nugget>();
		nuggets.add(new TestNugget(101, 0, 1));
		nuggets.add(new TestNugget(102, 0, 2));
		nuggets.add(new TestNugget(103, 0, 3));
		nuggets.add(new TestNugget(104, 0, 4));
		nuggets.add(new TestNugget(105, 0, 5));

		NuggetGroupManager nm = new NuggetGroupManager(nuggets);
		Set<NuggetGroupItem> itemsToSave = nm.doAddGroup(3, null);
		Iterator<Nugget> iter = nm.getNuggets().iterator();
		assertNugget(101, 0, 1, iter.next());
		assertNugget(102, 0, 2, iter.next());
		assertNugget(103, 0, 3, iter.next());
		assertNugget(104, 1, 4, iter.next());
		assertNugget(105, 1, 5, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 104, 105);

		itemsToSave = nm.doAddGroup(1, null);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 0, 1, iter.next());
		assertNugget(102, 1, 2, iter.next());
		assertNugget(103, 1, 3, iter.next());
		assertNugget(104, 2, 4, iter.next());
		assertNugget(105, 2, 5, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 102, 103, 104, 105);

		itemsToSave = nm.doAddGroup(0, null);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, 1, iter.next());
		assertNugget(102, 2, 2, iter.next());
		assertNugget(103, 2, 3, iter.next());
		assertNugget(104, 3, 4, iter.next());
		assertNugget(105, 3, 5, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 101, 102, 103, 104, 105);

		itemsToSave = nm.doAddGroup(7, null);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, 1, iter.next());
		assertNugget(102, 2, 2, iter.next());
		assertNugget(103, 2, 3, iter.next());
		assertNugget(104, 3, 4, iter.next());
		assertNugget(105, 4, 5, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 105);

		itemsToSave = nm.doAddGroup(9, null);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, 1, iter.next());
		assertNugget(102, 2, 2, iter.next());
		assertNugget(103, 2, 3, iter.next());
		assertNugget(104, 3, 4, iter.next());
		assertNugget(105, 4, 5, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave);

		try {
			nm.doAddGroup(11, null);
			fail("IndexOutOfBoundsException expected");
		} catch (IndexOutOfBoundsException e) {}
	}

	public void testRemoveGroup() {
		List<Nugget> nuggets = new ArrayList<Nugget>();
		nuggets.add(new TestNugget(101, 1, 1));
		nuggets.add(new TestNugget(102, 2, 2));
		nuggets.add(new TestNugget(103, 2, 3));
		nuggets.add(new TestNugget(104, 3, 4));
		nuggets.add(new TestNugget(105, 4, 5));

		NuggetGroupManager nm = new NuggetGroupManager(nuggets);
		try {
			nm.doRemoveGroup(9);
			fail("IndexOutOfBoundsException expected");
		} catch (IndexOutOfBoundsException e) {}

		try {
			nm.doRemoveGroup(8);
			fail("IllegalArgumentException expected");
		} catch (IllegalArgumentException e) {}

		Set<NuggetGroupItem> itemsToSave = nm.doRemoveGroup(5);
		Iterator<Nugget> iter = nm.getNuggets().iterator();
		assertNugget(101, 1, 1, iter.next());
		assertNugget(102, 2, 2, iter.next());
		assertNugget(103, 2, 3, iter.next());
		assertNugget(104, 2, 4, iter.next());
		assertNugget(105, 3, 5, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 104, 105);

		itemsToSave = nm.doRemoveGroup(0);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 0, 1, iter.next());
		assertNugget(102, 1, 2, iter.next());
		assertNugget(103, 1, 3, iter.next());
		assertNugget(104, 1, 4, iter.next());
		assertNugget(105, 2, 5, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 101, 102, 103, 104, 105);

		itemsToSave = nm.doRemoveGroup(1);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 0, 1, iter.next());
		assertNugget(102, 0, 2, iter.next());
		assertNugget(103, 0, 3, iter.next());
		assertNugget(104, 0, 4, iter.next());
		assertNugget(105, 1, 5, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 102, 103, 104, 105);

		itemsToSave = nm.doRemoveGroup(4);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 0, 1, iter.next());
		assertNugget(102, 0, 2, iter.next());
		assertNugget(103, 0, 3, iter.next());
		assertNugget(104, 0, 4, iter.next());
		assertNugget(105, 0, 5, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 105);
	}

	public void testMoveGroup() {
		List<Nugget> nuggets = new ArrayList<Nugget>();
		nuggets.add(new TestNugget(101, 1, 1));
		nuggets.add(new TestNugget(102, 2, 2));
		nuggets.add(new TestNugget(103, 2, 3));
		nuggets.add(new TestNugget(104, 2, 4));
		nuggets.add(new TestNugget(105, 3, 5));
		nuggets.add(new TestNugget(106, 3, 6));
		nuggets.add(new TestNugget(107, 4, 7));
		nuggets.add(new TestNugget(108, 4, 8));

		NuggetGroupManager nm = new NuggetGroupManager(nuggets);
		Set<NuggetGroupItem> itemsToSave = nm.doMoveGroup(6, 0);
		Iterator<Nugget> iter = nm.getNuggets().iterator();
		assertNugget(101, 2, 1, iter.next());
		assertNugget(102, 3, 2, iter.next());
		assertNugget(103, 3, 3, iter.next());
		assertNugget(104, 3, 4, iter.next());
		assertNugget(105, 3, 5, iter.next());
		assertNugget(106, 3, 6, iter.next());
		assertNugget(107, 4, 7, iter.next());
		assertNugget(108, 4, 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 101, 102, 103, 104);

		itemsToSave = nm.doMoveGroup(1, 4);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, 1, iter.next());
		assertNugget(102, 3, 2, iter.next());
		assertNugget(103, 3, 3, iter.next());
		assertNugget(104, 3, 4, iter.next());
		assertNugget(105, 3, 5, iter.next());
		assertNugget(106, 3, 6, iter.next());
		assertNugget(107, 4, 7, iter.next());
		assertNugget(108, 4, 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 101);

		itemsToSave = nm.doMoveGroup(3, 5);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, 1, iter.next());
		assertNugget(102, 2, 2, iter.next());
		assertNugget(103, 3, 3, iter.next());
		assertNugget(104, 3, 4, iter.next());
		assertNugget(105, 3, 5, iter.next());
		assertNugget(106, 3, 6, iter.next());
		assertNugget(107, 4, 7, iter.next());
		assertNugget(108, 4, 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 102);

		itemsToSave = nm.doMoveGroup(0, 12);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 0, 1, iter.next());
		assertNugget(102, 1, 2, iter.next());
		assertNugget(103, 2, 3, iter.next());
		assertNugget(104, 2, 4, iter.next());
		assertNugget(105, 2, 5, iter.next());
		assertNugget(106, 2, 6, iter.next());
		assertNugget(107, 3, 7, iter.next());
		assertNugget(108, 3, 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 101, 102, 103, 104, 105, 106, 107, 108);

		itemsToSave = nm.doMoveGroup(8, 6);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 0, 1, iter.next());
		assertNugget(102, 1, 2, iter.next());
		assertNugget(103, 2, 3, iter.next());
		assertNugget(104, 2, 4, iter.next());
		assertNugget(105, 3, 5, iter.next());
		assertNugget(106, 3, 6, iter.next());
		assertNugget(107, 3, 7, iter.next());
		assertNugget(108, 3, 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 105, 106);
	}

	public void testMoveNugget() {
		List<Nugget> nuggets = new ArrayList<Nugget>();
		nuggets.add(new TestNugget(101, 1, 1));
		nuggets.add(new TestNugget(102, 2, 2));
		nuggets.add(new TestNugget(103, 2, 3));
		nuggets.add(new TestNugget(104, 2, 4));
		nuggets.add(new TestNugget(105, 3, 5));
		nuggets.add(new TestNugget(106, 3, 6));
		nuggets.add(new TestNugget(107, 4, 7));
		nuggets.add(new TestNugget(108, 4, 8));

		NuggetGroupManager nm = new NuggetGroupManager(nuggets);
		Set<NuggetGroupItem> itemsToSave = nm.doMoveNugget(3, 6);
		Iterator<Nugget> iter = nm.getNuggets().iterator();
		assertNugget(101, 1, 1, iter.next());
		assertNugget(103, 2, 2, iter.next());
		assertNugget(104, 2, 3, iter.next());
		assertNugget(102, 2, 4, iter.next());
		assertNugget(105, 3, 5, iter.next());
		assertNugget(106, 3, 6, iter.next());
		assertNugget(107, 4, 7, iter.next());
		assertNugget(108, 4, 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 103, 104, 102);

		itemsToSave = nm.doMoveNugget(5, 3);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, 1, iter.next());
		assertNugget(102, 2, 2, iter.next());
		assertNugget(103, 2, 3, iter.next());
		assertNugget(104, 2, 4, iter.next());
		assertNugget(105, 3, 5, iter.next());
		assertNugget(106, 3, 6, iter.next());
		assertNugget(107, 4, 7, iter.next());
		assertNugget(108, 4, 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 102, 103, 104);

		itemsToSave = nm.doMoveNugget(3, 2);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, 1, iter.next());
		assertNugget(102, 1, 2, iter.next());
		assertNugget(103, 2, 3, iter.next());
		assertNugget(104, 2, 4, iter.next());
		assertNugget(105, 3, 5, iter.next());
		assertNugget(106, 3, 6, iter.next());
		assertNugget(107, 4, 7, iter.next());
		assertNugget(108, 4, 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 102);

		itemsToSave = nm.doMoveNugget(4, 2);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, 1, iter.next());
		assertNugget(103, 1, 2, iter.next());
		assertNugget(102, 1, 3, iter.next());
		assertNugget(104, 2, 4, iter.next());
		assertNugget(105, 3, 5, iter.next());
		assertNugget(106, 3, 6, iter.next());
		assertNugget(107, 4, 7, iter.next());
		assertNugget(108, 4, 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 103, 102);

		itemsToSave = nm.doMoveNugget(5, 10);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, 1, iter.next());
		assertNugget(103, 1, 2, iter.next());
		assertNugget(102, 1, 3, iter.next());
		assertNugget(105, 3, 4, iter.next());
		assertNugget(106, 3, 5, iter.next());
		assertNugget(104, 4, 6, iter.next());
		assertNugget(107, 4, 7, iter.next());
		assertNugget(108, 4, 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 105, 106, 104);

		itemsToSave = nm.doMoveNugget(6, 0);
		iter = nm.getNuggets().iterator();
		assertNugget(105, 0, 1, iter.next());
		assertNugget(101, 1, 2, iter.next());
		assertNugget(103, 1, 3, iter.next());
		assertNugget(102, 1, 4, iter.next());
		assertNugget(106, 3, 5, iter.next());
		assertNugget(104, 4, 6, iter.next());
		assertNugget(107, 4, 7, iter.next());
		assertNugget(108, 4, 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 105, 101, 102, 103);

		itemsToSave = nm.doMoveNugget(7, 12);
		iter = nm.getNuggets().iterator();
		assertNugget(105, 0, 1, iter.next());
		assertNugget(101, 1, 2, iter.next());
		assertNugget(103, 1, 3, iter.next());
		assertNugget(102, 1, 4, iter.next());
		assertNugget(104, 4, 5, iter.next());
		assertNugget(107, 4, 6, iter.next());
		assertNugget(108, 4, 7, iter.next());
		assertNugget(106, 4, 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 104, 107, 108, 106);
	}

	public void testMoveItem() {
		List<Nugget> nuggets = new ArrayList<Nugget>();
		nuggets.add(new TestNugget(101, 0, 1));
		nuggets.add(new TestNugget(102, 0, 2));
		nuggets.add(new TestNugget(103, 1, 3));
		nuggets.add(new TestNugget(104, 2, 4));
		nuggets.add(new TestNugget(105, 2, 5));

		NuggetGroupManager nm = new NuggetGroupManager(nuggets);
		try {
			nm.doMoveItem(-1, 0);
			fail("ArrayIndexOutOfBoundsException expected");
		} catch (ArrayIndexOutOfBoundsException e) {}
		try {
			nm.doMoveItem(7, 0);
			fail("IndexOutOfBoundsException expected");
		} catch (IndexOutOfBoundsException e) {}

		try {
			nm.doMoveItem(0, -1);
			fail("IndexOutOfBoundsException expected");
		} catch (IndexOutOfBoundsException e) {}
		try {
			nm.doMoveItem(0, 8);
			fail("IndexOutOfBoundsException expected");
		} catch (IndexOutOfBoundsException e) {}
	}

	public void testMakeConsistentBestNuggets() {
		List<Nugget> nuggets = new ArrayList<Nugget>();
		nuggets.add(new TestNugget(101, 0, 1, false));
		nuggets.add(new TestNugget(102, 1, 2, false));
		nuggets.add(new TestNugget(103, 1, 3, false));
		nuggets.add(new TestNugget(104, 2, 4, false));
		nuggets.add(new TestNugget(105, 2, 5, false));
		nuggets.add(new TestNugget(106, 3, 6, false));

		NuggetGroupManager nm = new NuggetGroupManager(nuggets);
		Set<NuggetGroupItem> itemsToSave = nm.doMakeConsistent();
		Iterator<Nugget> iter = nm.getNuggets().iterator();
		assertNugget(101, 0, 1, false, iter.next());
		assertNugget(102, 1, 2, false, iter.next());
		assertNugget(103, 1, 3, false, iter.next());
		assertNugget(104, 2, 4, false, iter.next());
		assertNugget(105, 2, 5, false, iter.next());
		assertNugget(106, 3, 6, false, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave);


		nuggets = new ArrayList<Nugget>();
		nuggets.add(new TestNugget(101, 0, 1, true));
		nuggets.add(new TestNugget(102, 1, 2, true));
		nuggets.add(new TestNugget(103, 1, 3, true));
		nuggets.add(new TestNugget(104, 2, 4, true));
		nuggets.add(new TestNugget(105, 2, 5, true));
		nuggets.add(new TestNugget(106, 3, 6, true));

		nm = new NuggetGroupManager(nuggets);
		itemsToSave = nm.doMakeConsistent();
		iter = nm.getNuggets().iterator();
		assertNugget(101, 0, 1, true, iter.next());
		assertNugget(102, 1, 2, true, iter.next());
		assertNugget(103, 1, 3, false, iter.next());
		assertNugget(104, 2, 4, true, iter.next());
		assertNugget(105, 2, 5, false, iter.next());
		assertNugget(106, 3, 6, true, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 103, 105);


		nuggets = new ArrayList<Nugget>();
		nuggets.add(new TestNugget(101, 0, 1, true));
		nuggets.add(new TestNugget(102, 1, 2, true));
		nuggets.add(new TestNugget(103, 1, 3, true));
		nuggets.add(new TestNugget(104, 3, 4, false));
		nuggets.add(new TestNugget(105, 3, 5, true));
		nuggets.add(new TestNugget(106, 4, 6, false));

		nm = new NuggetGroupManager(nuggets);
		itemsToSave = nm.doMakeConsistent();
		iter = nm.getNuggets().iterator();
		assertNugget(101, 0, 1, true, iter.next());
		assertNugget(102, 1, 2, true, iter.next());
		assertNugget(103, 1, 3, false, iter.next());
		assertNugget(104, 2, 4, false, iter.next());
		assertNugget(105, 2, 5, true, iter.next());
		assertNugget(106, 3, 6, false, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 103, 104, 105, 106);

		itemsToSave = nm.doMakeConsistent();
		iter = nm.getNuggets().iterator();
		assertNugget(101, 0, 1, true, iter.next());
		assertNugget(102, 1, 2, true, iter.next());
		assertNugget(103, 1, 3, false, iter.next());
		assertNugget(104, 2, 4, false, iter.next());
		assertNugget(105, 2, 5, true, iter.next());
		assertNugget(106, 3, 6, false, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave);


		nuggets = new ArrayList<Nugget>();
		nuggets.add(new TestNugget(101, 2, 2));
		nuggets.add(new TestNugget(102, 1, 3));
		nuggets.add(new TestNugget(103, 1, 5));
		nuggets.add(new TestNugget(104, 1, 5));
		nuggets.add(new TestNugget(105, 3, 5));
		nuggets.add(new TestNugget(106, 1, 9));

		nm = new NuggetGroupManager(nuggets);
		itemsToSave = nm.doMakeConsistent();
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, 1, iter.next());
		assertNugget(102, 2, 2, iter.next());
		assertNugget(103, 2, 3, iter.next());
		assertNugget(104, 2, 4, iter.next());
		assertNugget(105, 3, 5, iter.next());
		assertNugget(106, 4, 6, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 101, 102, 103, 104, 106);

		itemsToSave = nm.doMakeConsistent();
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, 1, iter.next());
		assertNugget(102, 2, 2, iter.next());
		assertNugget(103, 2, 3, iter.next());
		assertNugget(104, 2, 4, iter.next());
		assertNugget(105, 3, 5, iter.next());
		assertNugget(106, 4, 6, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave);
	}

	public void testSelectBestNugget() {
		List<Nugget> nuggets = new ArrayList<Nugget>();
		nuggets.add(new TestNugget(101, 1, 1, false));
		nuggets.add(new TestNugget(102, 1, 2, false));
		nuggets.add(new TestNugget(103, 1, 3, false));
		nuggets.add(new TestNugget(104, 2, 4, false));
		nuggets.add(new TestNugget(105, 2, 5, false));
		nuggets.add(new TestNugget(106, 3, 6, false));

		NuggetGroupManager nm = new NuggetGroupManager(nuggets);
		Set<NuggetGroupItem> itemsToSave = nm.doSelectBestNugget(0, 101);
		Iterator<Nugget> iter = nm.getNuggets().iterator();
		assertNugget(101, 1, 1, false, iter.next());
		assertNugget(102, 1, 2, false, iter.next());
		assertNugget(103, 1, 3, false, iter.next());
		assertNugget(104, 2, 4, false, iter.next());
		assertNugget(105, 2, 5, false, iter.next());
		assertNugget(106, 3, 6, false, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave);

		nm = new NuggetGroupManager(nuggets);
		itemsToSave = nm.doSelectBestNugget(1, 102);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, 1, false, iter.next());
		assertNugget(102, 1, 2, true, iter.next());
		assertNugget(103, 1, 3, false, iter.next());
		assertNugget(104, 2, 4, false, iter.next());
		assertNugget(105, 2, 5, false, iter.next());
		assertNugget(106, 3, 6, false, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 102);

		nm = new NuggetGroupManager(nuggets);
		itemsToSave = nm.doSelectBestNugget(1, 103);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, 1, false, iter.next());
		assertNugget(102, 1, 2, false, iter.next());
		assertNugget(103, 1, 3, true, iter.next());
		assertNugget(104, 2, 4, false, iter.next());
		assertNugget(105, 2, 5, false, iter.next());
		assertNugget(106, 3, 6, false, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 102, 103);

		nm = new NuggetGroupManager(nuggets);
		itemsToSave = nm.doSelectBestNugget(2, 104);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, 1, false, iter.next());
		assertNugget(102, 1, 2, false, iter.next());
		assertNugget(103, 1, 3, true, iter.next());
		assertNugget(104, 2, 4, true, iter.next());
		assertNugget(105, 2, 5, false, iter.next());
		assertNugget(106, 3, 6, false, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 104);

		nm = new NuggetGroupManager(nuggets);
		itemsToSave = nm.doSelectBestNugget(3, 106);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, 1, false, iter.next());
		assertNugget(102, 1, 2, false, iter.next());
		assertNugget(103, 1, 3, true, iter.next());
		assertNugget(104, 2, 4, true, iter.next());
		assertNugget(105, 2, 5, false, iter.next());
		assertNugget(106, 3, 6, true, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 106);

		nm = new NuggetGroupManager(nuggets);
		itemsToSave = nm.doSelectBestNugget(2, 105);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, 1, false, iter.next());
		assertNugget(102, 1, 2, false, iter.next());
		assertNugget(103, 1, 3, true, iter.next());
		assertNugget(104, 2, 4, false, iter.next());
		assertNugget(105, 2, 5, true, iter.next());
		assertNugget(106, 3, 6, true, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 104, 105);

		nm = new NuggetGroupManager(nuggets);
		itemsToSave = nm.doSelectBestNugget(2, 105);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, 1, false, iter.next());
		assertNugget(102, 1, 2, false, iter.next());
		assertNugget(103, 1, 3, true, iter.next());
		assertNugget(104, 2, 4, false, iter.next());
		assertNugget(105, 2, 5, true, iter.next());
		assertNugget(106, 3, 6, true, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave);

		nm = new NuggetGroupManager(nuggets);
		itemsToSave = nm.doSelectBestNugget(3, 105);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, 1, false, iter.next());
		assertNugget(102, 1, 2, false, iter.next());
		assertNugget(103, 1, 3, true, iter.next());
		assertNugget(104, 2, 4, false, iter.next());
		assertNugget(105, 2, 5, true, iter.next());
		assertNugget(106, 3, 6, false, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 106);

		nm = new NuggetGroupManager(nuggets);
		itemsToSave = nm.doSelectBestNugget(1, -1);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, 1, false, iter.next());
		assertNugget(102, 1, 2, false, iter.next());
		assertNugget(103, 1, 3, false, iter.next());
		assertNugget(104, 2, 4, false, iter.next());
		assertNugget(105, 2, 5, true, iter.next());
		assertNugget(106, 3, 6, false, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 103);
	}


	protected void assertNugget(final int id, final int group,
			final int orderPos, final Nugget actual) {
		assertEquals(id, actual.getId());
		assertEquals(group, actual.getGroup());
		assertEquals(orderPos, actual.getOrderPos());
	}

	protected void assertNugget(final int id, final int group,
			final int orderPos, final boolean bestNugget,
			final Nugget actual) {
		assertEquals(id, actual.getId());
		assertEquals(group, actual.getGroup());
		assertEquals(orderPos, actual.getOrderPos());
		assertEquals(bestNugget, actual.isBestNugget());
	}

	protected void assertChanges(final Set<NuggetGroupItem> itemsToSave,
			final int... expectedIDs) {
		TreeSet<Integer> expected = new TreeSet<Integer>();
		for (int expectedID : expectedIDs)
			expected.add(expectedID);

		TreeSet<Integer> actual = new TreeSet<Integer>();
		for (NuggetGroupItem item : itemsToSave)
			if (!item.isGroupHeader())
				actual.add(item.getNugget().getId());

		assertEquals(Arrays.toString(expected.toArray()),
				Arrays.toString(actual.toArray()));
	}

}
